+++ api/league/cust_league_sessions.go
@@ -204,19 +204,19 @@ type LeagueCustLeagueSessionsResponse struct {
 		AllowedTeams   []interface{} `json:"allowed_teams"`
 		AllowedLeagues []int         `json:"allowed_leagues"`
 		Cars           []struct {
-			CarId             int    `json:"car_id"`
-			CarName           string `json:"car_name"`
-			CarClassId        int    `json:"car_class_id"`
-			CarClassName      string `json:"car_class_name"`
-			MaxPctFuelFill    int    `json:"max_pct_fuel_fill"`
-			WeightPenaltyKg   int    `json:"weight_penalty_kg"`
-			PowerAdjustPct    int    `json:"power_adjust_pct"`
-			MaxDryTireSets    int    `json:"max_dry_tire_sets"`
-			QualSetupId       int    `json:"qual_setup_id"`
-			QualSetupFilename string `json:"qual_setup_filename"`
-			RaceSetupId       int    `json:"race_setup_id"`
-			RaceSetupFilename string `json:"race_setup_filename"`
-			PackageId         int    `json:"package_id"`
+			CarId             int     `json:"car_id"`
+			CarName           string  `json:"car_name"`
+			CarClassId        int     `json:"car_class_id"`
+			CarClassName      string  `json:"car_class_name"`
+			MaxPctFuelFill    float64 `json:"max_pct_fuel_fill"`
+			WeightPenaltyKg   int     `json:"weight_penalty_kg"`
+			PowerAdjustPct    float64 `json:"power_adjust_pct"`
+			MaxDryTireSets    int     `json:"max_dry_tire_sets"`
+			QualSetupId       int     `json:"qual_setup_id"`
+			QualSetupFilename string  `json:"qual_setup_filename"`
+			RaceSetupId       int     `json:"race_setup_id"`
+			RaceSetupFilename string  `json:"race_setup_filename"`
+			PackageId         int     `json:"package_id"`
 		} `json:"cars"`
 		HeatSesInfo struct {
 			ConsolationDeltaMaxFieldSize         int    `json:"consolation_delta_max_field_size"`
